Update Auth Decorators
